backend/requirements.txt
fastapi
uvicorn
sqlmodel
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt
asyncpg
psycopg2-binary 
python-multipart
email-validator
annotated-types
pydantic-core
python-dotenv
typing-extensions
typing-inspect
openai-agents
mcp
nest_asyncio
